requirements.txt
flask==3.1.2
flask-cors==6.0.2
torch>=2.5.0
torchvision>=0.20.0
transformers>=4.57.0
pillow>=12.0.0
numpy>=2.4.0
opencv-python>=4.9.0
matplotlib>=3.8.0
huggingface-hub>=0.36.0

